feat(rules): add no-important rule
Add rule to discourage use of !important in CSS declarations.
This is part of the @eslint/css parity implementation and is
a recommended rule that helps maintain clean CSS architecture.

The rule reports all instances of !important usage to encourage
developers to use proper CSS specificity instead.

Part of #eslint-css-parity

Signed-off-by: Alejandro Mery <amery@apptly.co>